TIP-842: Move attribute option value model to structure sub context
