+++ src/Akeneo/Pim/Structure/Bundle/DependencyInjection/Compiler/ResolveDoctrineTargetModelPass.php
@@ -10,6 +10,7 @@ use Akeneo\Pim\Structure\Component\Model\AttributeGroupInterface;
 use Akeneo\Pim\Structure\Component\Model\AttributeGroupTranslationInterface;
 use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
 use Akeneo\Pim\Structure\Component\Model\AttributeOptionInterface;
+use Akeneo\Pim\Structure\Component\Model\AttributeOptionValueInterface;
 use Akeneo\Pim\Structure\Component\Model\AttributeRequirementInterface;
 use Akeneo\Pim\Structure\Component\Model\AttributeTranslationInterface;
 use Akeneo\Pim\Structure\Component\Model\FamilyInterface;
@@ -51,6 +52,7 @@ class ResolveDoctrineTargetModelPass extends AbstractResolveDoctrineTargetModelP
             GroupTypeTranslationInterface::class => 'pim_catalog.entity.group_type_translation.class',
             AssociationTypeTranslationInterface::class => 'pim_catalog.entity.association_type_translation.class',
             AssociationTypeInterface::class => 'pim_catalog.entity.association_type.class',
+            AttributeOptionValueInterface::class => 'pim_catalog.entity.attribute_option_value.class',
         ];
     }
 }
