+++ src/Akeneo/Pim/Structure/Bundle/Resources/config/entities.yml
@@ -27,3 +27,5 @@ parameters:
     pim_catalog.entity.association_type.class: Akeneo\Pim\Structure\Component\Model\AssociationType
     pim_catalog.entity.association_type_translation.class: Akeneo\Pim\Structure\Component\Model\AssociationTypeTranslation
     pim_catalog.model.association_type_translation.interface: Akeneo\Pim\Structure\Component\Model\AssociationTypeTranslationInterface
+    pim_catalog.entity.attribute_option_value.class: Akeneo\Pim\Structure\Component\Model\AttributeOptionValue
+    pim_catalog.model.attribute_option_value.interface: Akeneo\Pim\Structure\Component\Model\AttributeOptionValueInterface
