+++ src/Akeneo/Pim/Structure/Component/spec/Normalizer/Versionning/AttributeNormalizerSpec.php
@@ -5,7 +5,7 @@ namespace spec\Akeneo\Pim\Structure\Component\Normalizer\Versionning;
 use Doctrine\Common\Collections\ArrayCollection;
 use PhpSpec\ObjectBehavior;
 use Akeneo\Pim\Structure\Component\Model\AttributeOption;
-use Pim\Bundle\CatalogBundle\Entity\AttributeOptionValue;
+use Akeneo\Pim\Structure\Component\Model\AttributeOptionValue;
 use Pim\Bundle\VersioningBundle\Normalizer\Flat\TranslationNormalizer;
 use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
 use Akeneo\Pim\Structure\Component\Normalizer\Standard\AttributeNormalizer;
