+++ src/Pim/Bundle/CatalogBundle/Resources/config/models.yml
@@ -6,7 +6,6 @@ parameters:
     pim_catalog.model.completeness.interface:           Pim\Component\Catalog\Model\CompletenessInterface
     pim_catalog.model.category.interface:               Pim\Component\Catalog\Model\CategoryInterface
     pim_catalog.model.group.interface:                  Pim\Component\Catalog\Model\GroupInterface
-    pim_catalog.model.attribute_option_value.interface: Pim\Component\Catalog\Model\AttributeOptionValueInterface
     pim_catalog.model.metric.interface:                 Pim\Component\Catalog\Model\MetricInterface
     pim_catalog.model.product_price.interface:          Pim\Component\Catalog\Model\ProductPriceInterface
     pim_catalog.model.category_translation.interface:         Pim\Component\Catalog\Model\CategoryTranslationInterface
