+++ src/Pim/Bundle/EnrichBundle/Normalizer/AttributeOptionValueCollectionNormalizer.php
@@ -2,8 +2,8 @@
 
 namespace Pim\Bundle\EnrichBundle\Normalizer;
 
+use Akeneo\Pim\Structure\Component\Model\AttributeOptionValueInterface;
 use Doctrine\Common\Collections\Collection;
-use Pim\Component\Catalog\Model\AttributeOptionValueInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 use Symfony\Component\Serializer\SerializerAwareInterface;
 use Symfony\Component\Serializer\SerializerInterface;
