+++ src/Pim/Bundle/EnrichBundle/spec/Form/Type/AttributeOptionValueTypeSpec.php
@@ -3,7 +3,7 @@
 namespace spec\Pim\Bundle\EnrichBundle\Form\Type;
 
 use PhpSpec\ObjectBehavior;
-use Pim\Bundle\CatalogBundle\Entity\AttributeOptionValue;
+use Akeneo\Pim\Structure\Component\Model\AttributeOptionValue;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\HiddenType;
 use Symfony\Component\Form\FormBuilderInterface;
