+++ src/Pim/Component/Catalog/spec/Value/OptionsValueSpec.php
@@ -5,7 +5,7 @@ namespace spec\Pim\Component\Catalog\Value;
 use PhpSpec\ObjectBehavior;
 use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
 use Akeneo\Pim\Structure\Component\Model\AttributeOptionInterface;
-use Pim\Component\Catalog\Model\AttributeOptionValueInterface;
+use Akeneo\Pim\Structure\Component\Model\AttributeOptionValueInterface;
 
 class OptionsValueSpec extends ObjectBehavior
 {
